### 2024-06-06 #6853 @editorialbot
[PRE REVIEW]: A tool for qualitative data analysis designed to support computational thinking
**Submitting author:** @cproctor (<a href="http://orcid.org/0000-0003-3492-9590">Chris Proctor</a>)
**Repository:** https://github.com/cproctor/qualitative-coding
**Branch with paper.md** (empty if default branch):
**Version:** 1.3.2
