### 2024-06-06 #6853 @crvernon
A note, the authors are using the following OSI approved license:  https://opensource.org/license/agpl-v3
