### 2024-06-10 #6853 @ymzayek
Hi @cproctor, thank you for submitting to JOSS. I am the handling editor for this submission. Do you have any suggestion for reviewers? You can use the [potential reviewers list](https://reviewers.joss.theoj.org/lookup) provided at the top of this issue. I think people interested in data analysis methods, the social sciences and computer science would be a good fit. If you have suggestions, please provide a list of potential reviewers here but do not use the @ character to avoid receiving unwanted notifications. Thank you in advance.
