### 2024-06-18 #6853 @cproctor
Hi @ymzayek, thanks for taking on this paper as editor. I read through the list of reviewers; the following appear to have relevant interests and expertise:

- [jhagerer](https://github.com/jhagerer) (NLP and social computing. Python)
- [hendr1km](https://github.com/hendr1km) (Educational research; psychology)
- [SamHames](https://github.com/SamHames) (Digital humanities; Python)
- [pmyteh](https://github.com/pmyteh) (Social science; text analysis; Python)
- [cmaimone](https://github.com/cmaimone) (Social science; text analysis; Python)
