### 2024-06-19 4baa217
Update README.md with JOSS badge
