### 2024-06-24 d2245f7
Add logo to docs page
