+++ docs/conf.py
@@ -26,3 +26,8 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 
 html_theme = 'sphinx_rtd_theme'
 html_static_path = ['_static']
+html_logo = '../qc_lockup.v0.png'
+html_theme_options = {
+    'logo_only': True,
+    'display_version': False,
+}
