### 2024-06-24 fad2dd9
Update license formatting
