### 2024-06-24 d812d89
Remove redundant top-level header in README
