### 2024-07-01 #6853 @cproctor
@ymzayek I know two invited reviewers have not yet written back but here are three more potential reviewers with relevant expertise, in case you want to invite others to review:

- [deniederhut](https://github.com/deniederhut) (Python, anthropology, natural language processing)
- [rmhogervorst](https://github.com/rmhogervorst) (Python, social science)
- [alexanderfurnas](https://github.com/alexanderfurnas) (Python, text analysis)

Thanks, everyone--I know everyone is busy.
