### 2024-07-02 #6853 @ymzayek
Hello @deniederhut and @rmhogervorst,

Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?

A tool for qualitative data analysis designed to support computational thinking

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/6853).

Thank you in advance!
