### 2024-07-05 #6853 @hendr1km
I am sorry, I do not have the capacity at the moment - although I will keep this software in mind for qualitative work in the future.
