### 2024-07-31 cproctor/qualitative-coding#45 @cmaimone
JOSS: Community Guidelines
https://github.com/openjournals/joss-reviews/issues/7031 - Community Guidelines

Suggestions:

- In the documentation https://qualitative-coding.readthedocs.io/en/latest/manuscript.html#contributing, link to https://www.contributor-covenant.org/
- Add a CONTRIBUTING file to the repo, so that people see the guidelines when going to the issue page

Needed:

- Clearer information on how someone contributes to the software, not just submit a bug report (do you want them to submit PRs?)
- Clearer information on getting support (create an issue?  email you?)
