### 2024-07-31 cproctor/qualitative-coding#45 @cproctor
Thanks! I appreciate this feedback, and should be able to address them shortly.
