### 2024-07-31 76513ae
Clearer information on contributing and support #45
