+++ CONTRIBUTING.md
@@ -1,0 +1,23 @@
+# Contributing
+
+[Chris Proctor](chrisp@buffalo.edu), the project lead, would be delighted to hear about your experience 
+using `qc`. Bug reports, feature requests, and discussion of the future directions of `qc` takes place 
+on the [project repository’s issues page](https://github.com/cproctor/qualitative-coding/issues).
+Code contributions to this project should be via pull requests on this repository.
+
+## How you can help
+
+`qc` is a young project; it's not yet clear how (or if) the project should grow to best support 
+its users. Probably the best way to contribute is to use `qc` in your own research, and to share
+limitations you discover or new features you wish existed. This 
+[repository’s issues page](https://github.com/cproctor/qualitative-coding/issues) would be a great
+place to iterate ideas for new features. 
+
+## Code of conduct
+
+[![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-2.1-4baaaa.svg)](code_of_conduct.md)
+
+This project adopts the [Contributor Covenant](https://www.contributor-covenant.org/). 
+Please contact [Chris Proctor](chrisp@buffalo.edu) if you experience treatment which makes you 
+feel unsafe or unwelcome. 
+
