+++ docs/manuscript.rst
@@ -854,13 +854,19 @@ Contributing
 
 Development of ``qc`` follows the Contributor Covenant. 
 
-.. |Contributor Covenant| image:: https://img.shields.io/badge/Contributor%20Covenant-2.1-4baaaa.svg :target: code_of_conduct.md
+.. image:: https://img.shields.io/badge/Contributor%20Covenant-2.1-4baaaa.svg 
+   :alt: Contributor Covenant 2.1 badge
+   :target: https://www.contributor-covenant.org/
 
 Chris Proctor (chrisp@buffalo.edu), the project lead, would be delighted to hear 
 about your experience using ``qc``.
 Bug reports, feature requests, and discussion of the future directions of 
 ``qc`` takes place on the project repository's
-`issues page <https://github.com/cproctor/qualitative-coding/issues>`__.
+`issues page <https://github.com/cproctor/qualitative-coding/issues>`__. 
+Code contributions to this project should be via pull requests on this repository. 
+
+If you are considering using ``qc`` in a research project or need help, you are
+also welcome to contact Chris directly via email. 
 
 Acknowledgements
 ================
