### 2024-07-31 b930d19
Update formatting of contributing section of docs
