### 2024-08-01 1518c46
Create migrations for 1.4.0, supporting logging
