### 2024-08-01 863853e
Refactor logs using structlog
