+++ qualitative_coding/cli/codes/rename.py
@@ -2,6 +2,7 @@ import click
 import os
 from qualitative_coding.corpus import QCCorpus
 from qualitative_coding.helpers import read_file_list
+from qualitative_coding.logs import configure_logger
 
 @click.command()
 @click.argument("old_codes", nargs=-1)
@@ -15,6 +16,9 @@ from qualitative_coding.helpers import read_file_list
 def rename(old_codes, new_code, settings, coders, pattern, filenames):
     "Rename one or more codes"
     settings_path = settings or os.environ.get("QC_SETTINGS", "settings.yaml")
+    log = configure_logger(settings_path)
+    log.info("rename", old_codes=old_codes, new_code=new_code, coders=coders, 
+             pattern=pattern, filenames=filenames)
     corpus = QCCorpus(settings_path)
     with corpus.session():
         corpus.rename_codes(
