+++ qualitative_coding/cli/init.py
@@ -1,6 +1,7 @@
 import click
 import os
 from qualitative_coding.cli.decorators import handle_qc_errors
+from qualitative_coding.logs import configure_logger
 from os import getcwd
 from pathlib import Path
 
@@ -13,6 +14,8 @@ from pathlib import Path
 def init(settings, accept_defaults, _import):
     "Initialize a qc project"
     settings_path = settings or os.environ.get("QC_SETTINGS", "settings.yaml")
+    log = configure_logger(settings_path)
+    log.info("init", accept_defaults=accept_defaults, _import=_import)
     if _import:
         from qualitative_coding.refi_qda.reader import REFIQDAReader
         reader = REFIQDAReader(_import)
