### 2024-08-01 cproctor/qualitative-coding#46 @cmaimone
Issue importing demo
Re: https://github.com/openjournals/joss-reviews/issues/7031

I followed the commands at https://qualitative-coding.readthedocs.io/en/latest/manuscript.html#exploring-an-existing-project but ran into an error when trying to init the project:

```
/Users/christina/Desktop/qc_demo % qc init --import ../demo.qdpx
../demo.qdpx is not a zipfile
```
