### 2024-08-01 cproctor/qualitative-coding#47 @cmaimone
Documentation issue
re: https://github.com/openjournals/joss-reviews/issues/7031

https://qualitative-coding.readthedocs.io/en/latest/manuscript.html#create-a-new-project -- has a command `qc import filename` but "import" does not seem to be a current command:

```
/Users/christina/Desktop/qc_project % qc init -y

/Users/christina/Desktop/qc_project % qc import ../untitled.txt
Usage: qc [OPTIONS] COMMAND [ARGS]...
Try 'qc --help' for help.

Error: No such command 'import'.

(base) /Users/christina/Desktop/qc_project % qc --help
Usage: qc [OPTIONS] COMMAND [ARGS]...

  Qualitative coding for computer scientists

Options:
  --help  Show this message and exit.

Commands:
  check          Check project for errors
  code           Open a file for coding
  codebook (cb)  Update the codebook
  coders         List all coders
  codes          Codes commands
  corpus         Corpus commands
  export         Export project as REFI-QDA
  init           Initialize a qc project
  memo           Write a memo
  upgrade        Upgrade project to new version of qc
  version        Show version number

```
