### 2024-08-01 cproctor/qualitative-coding#47 @cproctor
So sorry! This was an error in the docs. The command should be `qc corpus import`, not `qc import`. I just fixed the issue.
