### 2024-08-01 cproctor/qualitative-coding#46 @cproctor
Thanks! I'm working now on reproducing this error, then will hopefully fix it.
