### 2024-08-01 cproctor/qualitative-coding#46 @cproctor
It's an issue with the `curl` command--the file that gets downloaded is actually a html file (change the extension to `.html` and you can open it in the browser). Working on the correct command to fetch the binary zip file.
