### 2024-08-01 8adfec0
Add documentation for export command
