### 2024-08-01 cproctor/qualitative-coding#46 @cproctor
I decided to host the demo project on an external URL rather than as part of the repo. I've updated the docs with the new URL for the demo project (https://computationalliteracies.net/people/chris/demo.qdpx), and tested the tutorial workflow to confirm that it's working. Thanks!
