### 2024-08-01 #7031 @cproctor
Thanks for catching the issues with `qc corpus import` in the tutorial; I have corrected the docs and tested the changes. No changes were needed to the software itself.
