### 2024-08-06 591b29c
Remove debug option from logs. Log level is debug when verbose.
