### 2024-08-08 cproctor/qualitative-coding#51 @cmaimone
Documentation inconsistency
```
(base) /Users/christina/Desktop/qc_demo % qc init --import ../demo.qdpx
(base) /Users/christina/Desktop/qc_demo % qc corpus list
admin-1.txt
board_2017.txt
```

Is different from output shown in documentation: https://qualitative-coding.readthedocs.io/en/latest/manuscript.html#exploring-an-existing-project

Subsequent output (such as from `qc codes stats --recursive-codes --recursive-counts --min 10`) is also different, likely because there's a file missing

re: https://github.com/openjournals/joss-reviews/issues/7031
