### 2024-08-08 cproctor/qualitative-coding#53 @cmaimone
Documentation suggestion
Suggestions, not review requirements:

Specify a language for your [code blocks ](https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html#directive-code-block)in the documentation to improve readability via syntax highlighting.  Many qualitative analysts are not very experienced with command line, so every little bit helps.  For example, something like the [Bash lexer](https://pygments.org/docs/lexers/#pygments.lexers.shell.BashLexer) might work to show the prompt vs. the command vs. the output more clearly

It may also help to do a pass through to check that all commands that should be typed start with the command prompt `$` and output is distinguished differently.  The only command I've noticed missing the prompt so far is the installation command `pipx install qualitative-coding`.  It may also be useful to state in the installation section and somewhere at the beginning of the vignette that `$` is the command line prompt and they shouldn't type it.

Just something that we've learned from working with novice command line users in the past.

re: https://github.com/openjournals/joss-reviews/issues/7031
