### 2024-08-08 cproctor/qualitative-coding#54 @cmaimone
Error - maybe with text editor?
Attempting to start a coding session.  I have one text document imported in my project.  I have Visual Studio Code installed on my system, but not running/open.  I get this error:

```
(base) /Users/christina/Desktop/qc_project % qc code christina2
An incomplete coding session exists. Run qc code --recover to recover this coding session or qc code --abandon to abandon it.
```

The very first time I ran this, I got some slight variation on the error message, but now I'm getting this, even if I switch the user name. Not sure 1) why I got the error originally, and 2) what to do now, since it seems stuck on this.

Relatedly, the documentation says I can use a different editor if I want, but I didn't see how to do this (I might have missed it).  For example, if I want to use BBEdit, or RStudio, how do I set that?

re: https://github.com/openjournals/joss-reviews/issues/7031
