### 2024-08-08 cproctor/qualitative-coding#51 @cproctor
Thanks! I have been working over the last couple of days to generate the correct demo project--was hoping I'd get to it before you got back to the review :/ I'll have the update in place shortly, synchronized with the documentation.
