### 2024-08-08 cproctor/qualitative-coding#52 @cproctor
Thanks--these are both thoughtful suggestions which I'll incorporate asap.
