### 2024-08-09 f84310e
Update docs to clarify command-line instructions
