+++ docs/manuscript.rst
@@ -189,9 +189,9 @@ also interested in comparing my coding with that of other team members.
 ``qc``\ ’s queries offer a powerful array of options. Consider the
 following query, whose complexity is not atypical in normal usage:
 
-::
+.. code-block:: console
 
-   $ qc codes stats equity participation --recursive-codes --recursive-counts \
+   % qc codes stats equity participation --recursive-codes --recursive-counts \
    --pattern round1 --min 5 --depth 2 --unit paragraph --format latex
 
 This query reports counts for the codes “equity,” “participation,” and
@@ -264,9 +264,9 @@ for Linux) which has Python 3.9 or higher installed. If you want to
 install ``qc`` globally on your system, the cleanest approaach is to use
 `pipx <https://pipx.pypa.io/stable/>`__.
 
-::
+.. code-block:: console
 
-   pipx install qualitative-coding
+   % pipx install qualitative-coding
 
 If your research project is already contained within a Python package
 and you want to install ``qc`` as a local dependency, simply add
@@ -287,20 +287,20 @@ to the district’s primary and secondard schools (Proctor, Bigman, and
 Blikstein 2019). The commands below create a directory on the Desktop,
 download the demo project, and unpack it into the directory.
 
-::
+.. code-block:: console
 
-   $ cd ~/Desktop
-   $ curl -O https://github.com/cproctor/qualitative-coding/blob/main/demo.qdpx
-   $ mkdir qc_demo
-   $ cd qc_demo
-   $ qc init --import ../demo.qdpx
+   % cd ~/Desktop
+   % curl -O https://github.com/cproctor/qualitative-coding/blob/main/demo.qdpx
+   % mkdir qc_demo
+   % cd qc_demo
+   % qc init --import ../demo.qdpx
 
 Now let’s use ``qc`` to look around this project. Which documents are
 included in the corpus?
 
-::
+.. code-block:: console
 
-   $ qc corpus list
+   % qc corpus list
    admin.txt
    board_member.txt
    teacher.txt
@@ -314,9 +314,9 @@ that specific code as well as a total for that code and all of its
 children. Finally, ``--min 10`` filters out codes used fewer than ten
 times.
 
-::
+.. code-block:: console
 
-   $ qc codes stats --recursive-codes --recursive-counts --min 10
+   % qc codes stats --recursive-codes --recursive-counts --min 10
    Code                              Count    Total
    ------------------------------  -------  -------
    rq1_definition                        5       55
@@ -344,9 +344,9 @@ recursive counts, this time using the short form of these flags
 (``-ra``). We also add the ``--by-document`` flag, to get a sense of
 differences across interviewees.
 
-::
+.. code-block:: console
 
-   $ qc codes stats rationales -ra --by-document
+   % qc codes stats rationales -ra --by-document
                                    admin.txt    board_member.txt    teacher.txt    Total
    -------------------------------  -----------  ------------------  -------------  -------
    rq1_definition:rationales                 12                  10              5       27
@@ -361,9 +361,9 @@ came up more in interviews with the administrator and the board member
 than with the teacher. To explore this pattern, we can we can view lines
 coded with “equity,” or any of its child codes.
 
-::
+.. code-block:: console
 
-   $ qc codes find equity -r
+   % qc codes find equity -r
 
    admin.txt (7)
    ================================================================================
@@ -399,7 +399,7 @@ flag again counts all subcodes. This time we omit ``-r``; including
[…]
