### 2024-08-09 5c62cd5
Update log output formatting and add logs tests
