### 2024-08-09 6480e8b
Add check for pandoc with helpful error message
