+++ docs/index.rst
@@ -29,6 +29,8 @@ need to be familiar with using a command line in order to use ``qc``.
 * A code editor. You should install 
   `Visual Studio Code <https://code.visualstudio.com/>`__, the default
   editor, unless you prefer a different editor such as emacs or vim.
+* The `Sync Scroll <https://marketplace.visualstudio.com/items?itemName=dqisme.sync-scroll>`__
+  extension for Visual Studio Code.
 
 Install with pip or pipx
 ~~~~~~~~~~~~~~~~~~~~~~~~
