### 2024-08-19 5ac0b91
Refactor internal apis
Two internal APIs changed: first, update_coded_lines and
other corpus methods now take document file paths rather than
Documents--we only ever pulled the file path off the doc anyway,
so this simplifies code and saves a query. These methods also
check for pathlib.Path and cast to strings, as required.

The second API change is that corpus methods now refer to corpus
paths with absolute or relative paths, rather than implicitly
using paths relative to the corpus dir.
