+++ qualitative_coding/cli/corpus/move.py
@@ -2,6 +2,7 @@ import click
 import os
 from qualitative_coding.corpus import QCCorpus
 from qualitative_coding.logs import configure_logger
+from qualitative_coding.cli.decorators import handle_qc_errors
 
 @click.command()
 @click.argument("target")
@@ -9,6 +10,7 @@ from qualitative_coding.logs import configure_logger
 @click.option("-s", "--settings", type=click.Path(exists=True), help="Settings file")
 @click.option("-r", "--recursive", is_flag=True, 
         help="Recursively import from directory")
+@handle_qc_errors
 def move(target, destination, settings, recursive):
     "Move a file in the corpus"
     settings_path = settings or os.environ.get("QC_SETTINGS", "settings.yaml")
