+++ qualitative_coding/cli/corpus/__init__.py
@@ -4,6 +4,7 @@ from qualitative_coding.cli.corpus.list import list_corpus_paths
 from qualitative_coding.cli.corpus.import_media import import_media
 from qualitative_coding.cli.corpus.move import move
 from qualitative_coding.cli.corpus.remove import remove
+from qualitative_coding.cli.corpus.update import update
 
 @click.group(name="corpus", cls=ClickAliasedGroup)
 def corpus_group():
@@ -13,3 +14,4 @@ corpus_group.add_command(list_corpus_paths, aliases=["ls"])
 corpus_group.add_command(move, aliases=["mv"])
 corpus_group.add_command(remove, aliases=["rm"])
 corpus_group.add_command(import_media)
+corpus_group.add_command(update)
