+++ qualitative_coding/cli/corpus/__init__.py
@@ -5,6 +5,7 @@ from qualitative_coding.cli.corpus.import_media import import_media
 from qualitative_coding.cli.corpus.move import move
 from qualitative_coding.cli.corpus.remove import remove
 from qualitative_coding.cli.corpus.update import update
+from qualitative_coding.cli.corpus.anonymize import anonymize
 
 @click.group(name="corpus", cls=ClickAliasedGroup)
 def corpus_group():
@@ -15,3 +16,4 @@ corpus_group.add_command(move, aliases=["mv"])
 corpus_group.add_command(remove, aliases=["rm"])
 corpus_group.add_command(import_media)
 corpus_group.add_command(update)
+corpus_group.add_command(anonymize)
