+++ pyproject.toml
@@ -13,7 +13,7 @@ include = ["refi_qda/schema.xml"]
 repository = "https://github.com/cproctor/qualitative-coding"
 
 [tool.poetry.dependencies]
-python = ">=3.9"
+python = ">=3.9,<4.0"
 tqdm = "^4.66.3"
 tabulate = "^0.9.0"
 numpy = "^1.26.0"
@@ -25,6 +25,7 @@ xmlschema = "^3.0.2"
 structlog = "^24.4.0"
 rich = "^13.7.1"
 more-itertools = "^10.4.0"
+spacy = "^3.7.5"
 
 [tool.poetry.group.docs.dependencies]
 sphinx = "^7.3.7"
