### 2024-08-19 d671aef
Add reverse option to anonymize
