### 2024-08-19 f181b1c
qc update updates file hash, add test
