### 2024-08-19 9c56d42
Remove needless debug logging
