### 2024-08-19 6fdf87b
Sync corpus rm api with other corpus apis
