### 2024-08-19 b7123b9
delete coded lines when removing doc
