### 2024-08-20 cproctor/qualitative-coding#52 @cproctor
I've adopted both suggestions in this thread. Thanks!
