### 2024-08-20 cproctor/qualitative-coding#53 @cproctor
Thanks! I didn't know about the rst lexers; I was able to achieve exactly the desired effect, with the shell prompt displayed but not selectable. I also added a note at the top of the documentation.
