### 2024-08-20 87a7ae9
Add note to manuscript about Sublime Text coding
