### 2024-08-20 3448ee2
Improve error msg when incomplete coding session exists
