### 2024-08-20 cproctor/qualitative-coding#54 @cproctor
I have substantially expanded the documentation in response to these thoughtful points:
- The prerequisites section now lists Sublime Text and its Scroll Sync extension, which is helpful.
- A note in the vignette explains how to align the document and the coding file when using VS Code.
- Regarding editor support, I have added a section documenting all the settings; the [section on editors](https://qualitative-coding.readthedocs.io/en/latest/manuscript.html#editor) explains how to specify additional editors.

In this specific situation, I'm guessing that maybe you ran `qc code` once before you had VS Code installed, so the invocation failed and left the project with an open coding session. I would expect that `qc code christina2 --abandon` would get you back to a clean state. Please let me know if the issue still appears.
