### 2024-08-20 #7031 @cproctor
I believe all open issues have now been addressed. When testing, please make sure you have version 1.4.1 or better. Thanks!
