### 2024-08-22 cproctor/qualitative-coding#55 @cmaimone
Crosstab counting inconsistency?
Working through the vignette, I ran a crosstab with two different units (default of line, and paragraph).  Why are the counts along the diagonal not the same?

```
(base) /Users/christina/Desktop/qc_demo % qc codes crosstab equity participation committee_participation -a
                   code    equity    participation    committee_participation
-----------------------  --------  ---------------  -------------------------
                 equity        15                7                          0
          participation                         15                          0
committee_participation                                                     9
(base) /Users/christina/Desktop/qc_demo % qc codes crosstab equity participation committee_participation -a --unit paragraph
                   code    equity    participation    committee_participation
-----------------------  --------  ---------------  -------------------------
                 equity        15                7                          0
          participation                         13                          0
committee_participation                                                     6
```

re: https://github.com/openjournals/joss-reviews/issues/7031
