### 2024-08-22 cproctor/qualitative-coding#56 @cmaimone
Documentation clarification: Line numbers, coding file?
Just as a note, as you indicate for Sublime Text, when VS Code opens for me, I have two tabs -- the files don't open side by side (I'm using a mac if it matters).  So it's not as intuitive that the values in the codes.txt file should go in the same line number as the lines in admin.txt.  The screenshot (Figure 1) helps, as the text itself wasn't completely clear that the line numbers need to match.

Even when the documents are side by side, they don't necessarily scroll together, so being very clear that the line numbers need to match is key.  Keeping the two files lined up is actually a bit challenging.

I was also confused by "an empty document for codes" - is this an entirely new document without any filename, or is it codes.txt?  I assumed codes.txt, but since there's only one such file in `qc_demo`, that doesn't seem right.   And when I looked at the code stats after, they didn't seem to reflect the codes, so I'm thinking maybe that's not the right file?

re: https://github.com/openjournals/joss-reviews/issues/7031
