### 2024-08-22 cproctor/qualitative-coding#57 @cmaimone
Modifying codebook hierarchy?
To create new codes or add hierarchical structure to the codebook, am I supposed to open and edit codebook.yaml directly?  That wasn't clear from the documentation.  There are commands for renaming and other changes to the codes, but not for "organizing the codebook"

re: https://github.com/openjournals/joss-reviews/issues/7031
