### 2024-08-22 cproctor/qualitative-coding#59 @cmaimone
How to end coding session?
When starting a coding session, the process in the console stays active after the editor is opened (you don't gain the command prompt back).  Is it OK to end this process in the console?  If not, what is the expected way to end a coding session?

re: https://github.com/openjournals/joss-reviews/issues/7031
