### 2024-08-22 cproctor/qualitative-coding#56 @cmaimone
For context, I ran opened a coding session in the qc_demo project to code the admin.txt file
