### 2024-08-23 cproctor/qualitative-coding#55 @cproctor
This is the intended behavior. If the same code is applied to multiple lines within a paragraph, these will be counted separately with a line-based unit of analysis, but will only count once when counting by paragraph. If this behavior is surprising, please let me know and I'll extend the docs to explain. Thanks!
