### 2024-08-23 2e27e60
remove redundancy in install instructions
