### 2024-08-23 cproctor/qualitative-coding#56 @cproctor
This is the intended behavior. The desired coding experience is to have the corpus file and the temporary coding file (`codes.txt`) side-by-side, scrolling together. Unfortunately, there isn't a way to invoke VS Code from the command to achieve this, so you need to manually drag them side-by-side. If you have the VS Code Sync Scroll Extension installed, the two side-by-side panes will scroll together.

The [installation instructions](https://qualitative-coding.readthedocs.io/en/latest/index.html#installation) are updated to include the Scroll Sync Extension, but I just realized that the vignette contained its own "Installation" section, which I had not updated. I removed the redundant installation instructions and linked to the correct instructions.

(The most graceful usage of `qc` is with vim, where you can get the desired coding behavior through a command-line invocation. I chose to introduce `qc` with VS Code in the vignette because many potential users will not know vim.)

I believe this issue can be closed, but please let me know if you think I should make any changes to make this clearer. Thanks!
