### 2024-08-23 cproctor/qualitative-coding#56 @cmaimone
It's more that I'm confused about how codes.txt is supposed to work.  The vignette at least I don't think specifically named the file (that it is called codes.txt), so that was some of the confusion.  But I'm also confused about how it works with there just being a single file.  What prompts an update to the database of codes after you edit that file?  The process stays open in the console once the files are opened in an editor, I'm not clear on the expected steps to close that out (other than killing the process).
