### 2024-08-23 cproctor/qualitative-coding#57 @cproctor
The codebook is automatically updated when you add new codes through coding. To add hierarchical structure, you just edit the codebook file directly. In case you accidentally delete codes from the codebook which are present in the corpus, you can run `qc codebook` to update the codebook, or you can just keep coding and missing codes will be added after you end your next coding session.

I agree this wasn't as clear as it could be in the vignette, so I just edited the language a bit. Some more detail is also given in the documentation for the [`qc codebook`](file:///Users/chrisp/Repos/qualitative-coding/docs/_build/html/manuscript.html#codebook-cb) command.

I think this issue can be closed, but I'll let you decide. Thanks!
