### 2024-08-23 cproctor/qualitative-coding#59 @cproctor
The editor is invoked using `--wait`, which keeps the process hanging until you save and close the files being edited. At that time, `qc` will read the coding file, update the codes, etc. I believe this is adequately described in the [docs](Save and close the codes document to end the coding session.) with "Save and close the codes document to end the coding session," but let me know if you have any thoughts for making it clearer without adding unnecessary detail to the vignette.

It's not a good idea to kill the process directly; this would be indistinguishable to `qc` from the editor crashing. Therefore, killing the process would leave you in an incomplete coding session and you'd need to use `qc code coder --recover` or `qc code coder --abandon` to clean up.

I believe this issue can be closed; let me know if you recommend changes. Thanks!
