### 2024-08-23 ed05f82
Add worked example of custom editor, explaining coding process.
