### 2024-08-23 cproctor/qualitative-coding#59 @cmaimone
In working through this workflow, I'm finding it very frustrating, as if I do the wrong thing (such as end the process in the console), I lose my work.  I had a coding session opened, and saved the codes.txt file, but then ended the console process.  I then restarted the coding process with `qc code` for the same file, and my work is gone.  I'm assuming that codes.txt gets generated each time from the database of codes, which is why this is happening -- because the file wasn't closed in the expected way, so an update to the database didn't happen.  But I had a codes.txt file I thought was saved with work in it, and it just gets deleted.

Related issue: https://github.com/cproctor/qualitative-coding/issues/60

I understand that the software is working as you designed, but fwiw, the workflow is fragile enough that I would be very hesitant to use it with my projects for real.  Have other people used qc?  Maybe it's just me.

As it's working as designed, this isn't an issue that blocks the JOSS review.
