### 2024-08-23 cproctor/qualitative-coding#61 @cmaimone
qc memo requires coder argument
https://qualitative-coding.readthedocs.io/en/latest/manuscript.html#memo - this isn't a valid example.  The command requires that the coder be specified.

I also can't list memos except for a specific coder - I can't see all of them?

```
% qc memo -l
Usage: qc memo [OPTIONS] CODER
Try 'qc memo --help' for help.

Error: Missing argument 'CODER'.
```

re: https://github.com/openjournals/joss-reviews/issues/7031
