### 2024-08-23 cproctor/qualitative-coding#61 @cproctor
I agree, it could be useful to see all memos together, however listing memos is currently implemented as a flag on the `memo` command (e.g. `qc memo chris --list`), so the coder is required. I could break listing out as a separate command (e.g. `qc memo list --coder chris` or `qc memo list`), but this would present ambiguity (e.g. is`qc memo list` a request to list memos or a request to write a memo with coder named "list"?

I could also add an `--all` flag to the current command, but it feels clumsy to specify a coder and then to essentially follow that with "oh wait actually show me everyone's memos."

As I don't see an elegant solution and the use case doesn't feel urgent, I'm inclined not to make a change. What do you think?
