### 2024-08-23 cproctor/qualitative-coding#61 @cmaimone
I'm not sure how you're using the memo files in practice, so I don't know what functionality would be best.  From an outside perspective, I would find it awkward to have to look at essentially individual memo files - I'd see that as more of just personal memos then instead of a team coordination tool.  Maybe that's what it's intended to be?
