### 2024-08-23 cproctor/qualitative-coding#62 @cmaimone
anonymize and spacy
I didn't see anywhere where the need for spacy was noted in the documentation - until I tried to use anonymize.  The message points to what is needed, but if someone doesn't have spacy already installed, then the suggestion won't work:

```
% qc corpus anonymize
A language model is required to run this task. Please run:
python -m spacy download en_core_web_sm

% python -m spacy download en_core_web_sm
/opt/anaconda3/bin/python: No module named spacy
```

re: https://github.com/openjournals/joss-reviews/issues/7031
