### 2024-08-23 cproctor/qualitative-coding#62 @cmaimone
I did install spacy and downloaded the language model, but I'm still getting the error message - probably because I installed qc via pipx and qc is running in a different environment than my base python installation.  I'm not sure actually how I'd add spacy to the environment that qc is running in
