### 2024-08-26 cproctor/qualitative-coding#62 @cproctor
The spacy library is installed during qc's installation, but the language model is not--spacy's authors [strongly discourage](https://github.com/explosion/spaCy/discussions/10608#discussioncomment-2498611) automatic installation of models, so I defer to them and require the user to install a language model manually when it is needed.

I had a similar issue of installing the model and then having it not found... here's a [debugging thread](https://stackoverflow.com/questions/54334304/spacy-cant-find-model-en-core-web-sm-on-windows-10-and-python-3-5-3-anacon), including for the case that appears to be yours, of running within conda: `conda install -c conda-forge spacy-model-en_core_web_sm`

Any luck?
