### 2024-08-27 cproctor/qualitative-coding#62 @cmaimone
No - I think the issue is that qc is installed via pipx, and that creates a whole separate Python environment.  So Python running in that environment doesn't know to look in the places that the spacy models would be when I install them outside of that environment, with pip or conda.  For example, when I install the language model, it ends up in: `/opt/anaconda3/lib/python3.11/site-packages/en_core_web_sm/en_core_web_sm-3.7.1` but it should probably be in something like `~/.local/pipx/venvs/qualitative-coding/lib/python3.11/site-packages/spacy/`.  So I think either you'd need to use an environment variable in qc to look for the library in a specific place (and also direct people to download it there), or I think you can make it a dependency in pyproject.toml, which would then get pipx to include it.  See for example https://github.com/explosion/spaCy/discussions/12399 or https://stackoverflow.com/questions/76314229/how-to-download-spacy-models-in-a-poetry-managed-environment
