### 2024-08-27 cproctor/qualitative-coding#62 @cproctor
Ooh, thanks! I will see if I can get this to work... I'd prefer not to have to prompt the user to download a model.

I think the idea would be for `qc` to lazily download the model when it's first needed. This would reduce the package size for users who don't use NLP features, and would also allow model specification in the settings file.

I should be able to provide a new release and follow up on open issues toward the end of the week... gotta get my [fall courses](https://computationalliteracies.net/courses/pedprog/) launched!
