### 2024-09-03 b9258b5
Fix several edge cases in handling error crashes
