+++ qualitative_coding/views/viewer.py
@@ -16,12 +16,15 @@ import csv
 import yaml
 import json
 import re
+import structlog
+
+log = structlog.get_logger()
 
 class QCCorpusViewer:
 
     codes_file = "codes.txt"
     coding_session_metadata_file = ".coding_session"
-    code_pattern = "[a-zA-Z0-9_\-]"
+    code_pattern = "^[a-zA-Z0-9_\-]+$"
 
     def __init__(self, corpus):
         self.corpus = corpus
@@ -594,32 +597,42 @@ class QCCorpusViewer:
         text = [f.read_text() for f in sorted(self.corpus.memos_dir.glob("*.md"))]
         return "\n\n".join(text)
 
-    def open_editor(self, corpus_file_path, coder_name):
+    def open_editor(self, corpus_file_path, coder_name, write_codes_file=True):
+        """Opens an editor for coding. 
+        corpus_file_path should be a string naming a path relative to the corpus_dir.
+        """
+        log.debug("Opening editor", corpus_file=corpus_file_path, coder=coder_name)
         codes_file_path = self.corpus.resolve_path(self.codes_file)
         full_path = self.corpus.corpus_dir / corpus_file_path
-        codes_file_path.write_text(self.codes_file_text(corpus_file_path, coder_name))
+        if write_codes_file:
+            codes_file_path.write_text(self.codes_file_text(corpus_file_path, coder_name))
         command = self.get_code_command(full_path, codes_file_path)
         try:
             p = run(command, shell=True, check=True)
             corpus_file_length = len(full_path.read_text().splitlines())
-            coded_lines = self.parse_codes(codes_file_path.read_text(), corpus_file_length)
+            coded_lines = self.parse_codes(coder_name, codes_file_path.read_text(), 
+                    corpus_file_length)
         except CalledProcessError as err:
             self.save_incomplete_coding_session(corpus_file_path, coder_name)
             raise QCError(
-                "The editor closed before coding was complete" + 
-                (f":\n{err.stderr}\n" if err.stderr else '') + 
-                "Run qc code --recover to recover this coding session or " + 
-                "qc code --abandon to abandon it."
+                "The editor closed before coding was complete, resulting in an " + 
+                "incomplete coding session" + 
+                (f":\n{err.stderr}. " if err.stderr else '. ') + 
+                f"Run qc code {coder_name} --recover to recover this coding session or " + 
+                f"qc code {coder_name} --abandon to abandon it."
             )
+        except CodeFileParseError as err:
+            self.save_incomplete_coding_session(corpus_file_path, coder_name)
+            raise err
 
         with self.corpus.session():
-            self.corpus.update_coded_lines(corpus_file_path, coder_name, coded_lines)
+            self.corpus.update_coded_lines(str(corpus_file_path), coder_name, coded_lines)
         codes_file_path.unlink()
         metadata_file_path = self.corpus.resolve_path(self.coding_session_metadata_file)
         if metadata_file_path.exists():
             metadata_file_path.unlink()
 
-    def parse_codes(self, code_file_text, expected_length):
+    def parse_codes(self, coder_name, code_file_text, expected_length):
         """Parses the contents of a code file after user editing. 
         If successful, returns coded_lines, a list of {line: int, code_id: str}
         Raises CodeFileParseError if there is an error parsing the file.
@@ -628,36 +641,49 @@ class QCCorpusViewer:
         if len(code_file_lines) != expected_length:
             raise CodeFileParseError(
                 "Expected the codes file to have the same number of lines as the document " + 
-                f"being coded ({expected_length}), but it has {len(code_file_lines)}."
+                f"being coded ({expected_length}), but it has {len(code_file_lines)}. " + 
[…]
