### 2024-09-03 a7cba5b
Bump to 1.4.2
