### 2024-09-03 cproctor/qualitative-coding#59 @cproctor
Thanks--I appreciate the feedback, and am especially grateful for your reviewing work, crashing into errors that I should have caught previously. I give more details in #60, but in short the intended workflow is less brittle than what you encountered. A coding session essentially involves editing `codes.txt`; once you save and close that file in your editor of choice, the qc process automatically continues, reading `codes.txt`, updating the database, and then deleting temporary files.

Previously, there were several bugs: it was possible to start a new coding session, overwriting existing work in `codes.txt`, and a few edge cases were not well-handled. I believe the process is now pretty resilient, comparable to the way git opens an editor to write a commit message.

qc is currently used quite a bit in my lab group, and by several collaborators. I'm aware of one other person who has used it in a published paper. It's admittedly a pretty specific tool, compared to GUI-based QDA tools like Atlas.TI, Dedoose, and Nvivo, but there's no reason to have unnecessary sharp edges. This review has helped move the project closer to my goal, of having a tool that works well in a specific niche (with some necessary prerequisite skills), without gatekeeping. Thanks!

I believe this issue can be closed.
