### 2024-09-05 ec658db
New strategy for installing spacy model
