### 2024-09-05 cproctor/qualitative-coding#62 @cproctor
I have just released 1.5.2, which uses the following strategy:
- The spacy language model is not a dependency and is not installed when the main qualitative-coding package is installed.
- qc attempts to automatically download and install the language model when `qc corpus anonymize` is invoked. The download strategy uses the spacy package's built-in functionality, maximizing compatibility. I have tested this in both a project virtual environment and the pipx environment.
- In case the automatic installation fails, the error is handled with a message telling the user they need to install the language model manually, and providing the recommended spacy command to do so.

After trying a variety of options, this approach feels best to me. Here are my rationales for rejecting alternatives:
- **Make the language model a regular dependency.** The small model is only 12mb, but I still prefer not to bloat the package size for an add-on that many people will not use. Also, I plan to extend the NLP/AI functionality of `qc` in the future, and this will probably involve allowing the user to select the local or remote language model they want to use. The more accurate spacy model, `en_core_web_trf`, is almost 500 mb, so automatically installing models will become less and less attractive.
- **Make the language model an [extra dependency](https://python-poetry.org/docs/pyproject/#extras)** specified in `pyproject.toml`. I tried to do this repeatedly, and couldn't find a way to get PyPI to accept the package. In the end, it feels like the wrong approach anyway, as I would be duplicating functionality (downloading and installing models) which is already provided by spacy, and which is properly the responsibility of spacy. I'm reluctant to take on the maintenance and support burden of this redundant functionality.

Does this feel like a reasonable approach? If so, I believe this issue can be closed. Thank you again for the work you've put into this review.
