### 2024-09-14 be4fa5f
Settings file need not exist for qc init
