### 2024-09-14 f05c416
qc init now creates project assets by default
