+++ docs/manuscript.rst
@@ -210,6 +210,8 @@ constructing SQL queries. This design is also informed by the UNIX
 philosophy (McIlroy, PInson, and Tague 1978), that programs should do
 one thing but do it well.
 
+.. _textfiles:
+
 Storing data in text files
 --------------------------
 
@@ -223,6 +225,8 @@ All files related to a ``qc`` project are contained within a directory:
    │   ├── admin.txt
    │   ├── board_member.txt
    │   └── teacher.txt
+   ├── memos
+   ├── qualitative_coding.log
    ├── qualitative_coding.sqlite3
    └── settings.yaml
 
@@ -239,7 +243,8 @@ a straightforward schema which could be integrated into other
 applications. (The initial version of ``qc`` stored all data in text
 files; the database was added to improve performance.) In almost all of
 our existing ``qc`` projects, QDA is part of an analytical pipeline. The
-prior and subsequent processes operate on
+prior and subsequent processes operate on these standard interfaces, 
+supporting flexible and creative uses of qualitative coding.
 
 Vignette
 ========
@@ -260,11 +265,11 @@ Installation
 
 Start by following the steps in :ref:`installation`. 
 
-Exploring an existing project
------------------------------
+Explore an existing project
+---------------------------
 
 The fastest way to experience what ``qc`` has to offer is to start with
-an exisitng project: an excerpt from the coded interview transcripts
+an existing project: an excerpt from the coded interview transcripts
 from members of a committee considering whether to add computer science
 to the district’s primary and secondard schools (Proctor, Bigman, and
 Blikstein 2019). The commands below create a directory on the Desktop,
@@ -418,15 +423,55 @@ other formats for inclusion in a publication.
 Create a new project
 --------------------
 
-This section is a guide to setting up a new ``qc`` project. Before
-starting, select one or more documents to add to your project’s corpus.
-These could be text documents in almost any file format (the default
+Setup
+~~~~~
+
+This section is a guide to starting a new ``qc`` project. Either save
+the excerpt below as ``teacher_2.txt`` on your Desktop (another teacher 
+interview from the project described in the previous section), or select 
+one or more of your own documents to work with.
+These could be in any text-based file format (the default
 importer uses `Pandoc <https://pandoc.org/>`__, and so can handle any
-file format supported by Pandoc.)
+file format supported by Pandoc.) 
+
+.. code-block:: console
+
+   Research_Assistant:         So just to start, it would be helpful to 
+   understand what your background is in computer science and also with 
+   Location 63.
+
+   Person_121:    So, I worked for years in the tech industry before I decided
+   to teach elementary school. I wasn't, so my background is engineering but not 
+   computer science, though the work I was doing was programming. And so after 
+   fifteen years in the tech industry, I lost my job, I went back to school, I 
+   did my multiple subject credential, and I started teaching as an elementary 
+   school teacher in Location 63.
+
+   Person_121:    So I literally just started as a classroom teacher, nothing 
+   else. I had no CS in mind at that time. And right about the time I started in 
+   Year was around the time Location 63 was getting smart boards in the classroom, 
+   one by one and I got to test that.  And then one thing led to another and I 
+   noticed how students who otherwise may not be motivated learning were motivated 
+   by the use of technology, so I got more into using technology with teaching to 
+   engage the students and engage them as a result.
+
+   Person_121:    During the course of that, I stumbled upon Scratch, and 
+   you're probably familiar with Scratch, I'm not sure, but it's the language 
+   developed at MIT Media Lab with children in mind. While I was still a fifth 
+   grade classroom teacher, I stumbled upon that and I thought oh this is 
[…]
