+++ paper.bib
@@ -8,10 +8,33 @@
   year={2022},
   publisher={Medical Library Association}
 }
+@article{engelbart1962,
+  title = {Augmenting Human Intellect: A Conceptual Framework},
+  author = {Engelbart, Douglas C},
+  year = {1962},
+  pages = {64--90}
+}
 @techreport{healy2020,
   title = {The {{Plain Person}}'s {{Guide}} to {{Plain Text Social Science}}},
   author = {Healy, Kieran},
   year = {2020}
+}
+@article{kidder1987,
+  title = {Qualitative and Quantitative Methods: {{When}} Stories Converge},
+  shorttitle = {Qualitative and Quantitative Methods},
+  author = {Kidder, Louise H. and Fine, Michelle},
+  year = {1987},
+  month = sep,
+  journal = {New Directions for Program Evaluation},
+  volume = {1987},
+  number = {35},
+  pages = {57--75},
+  issn = {0164-7989, 1551-2371},
+  doi = {10.1002/ev.1459},
+  urldate = {2024-09-15},
+  abstract = {Abstract             The use of qualitative measures in a quantitative framework results in a reasonable likelihood of triangulation; in contrast, the independent conduct of qualitative and quantitative evaluations is a greater challenge for triangulation, but it also holds promise for greater discovery.},
+  copyright = {http://onlinelibrary.wiley.com/termsAndConditions\#vor},
+  langid = {english}
 }
 @article{kuckartz2010realizing,
   title={Realizing mixed-methods approaches with MAXQDA},
@@ -30,6 +53,21 @@
   title={The Delphi technique: A credible research methodology},
   author={Ogbeifun, E and Agwa-Ejon, J and Mbohwa, Charles and Pretorius, JH},
   year={2016}
+}
+@incollection{pea1997,
+  title = {Practices of Distributed Intelligence and Designs for Education},
+  booktitle = {Distributed Cognitions: Psychological and Educational Considerations},
+  author = {Pea, Roy D},
+  editor = {Salomon, Gavriel},
+  year = {1997},
+  series = {Learning in Doing},
+  pages = {47--87},
+  publisher = {Cambridge University Press},
+  address = {New York},
+  isbn = {978-0-521-57423-5 978-0-521-41406-7},
+  langid = {english},
+  lccn = {BF311 .D538 1997},
+  keywords = {{Learning, Psychology of},Distributed cognition,Social aspects}
 }
 @inproceedings{proctor2019,
   title = {Defining and Designing Computer Science Education in a {{K12}} Public School District},
@@ -65,4 +103,12 @@
   pages={65--75},
   year={2002},
   publisher={University of the Free State}
+}
+@article{wing2011research,
+  title = {Research Notebook: {{Computational}} Thinking---{{What}} and Why},
+  author = {Wing, Jeanette},
+  year = {2011},
+  journal = {The link magazine},
+  volume = {6},
+  pages = {20--23}
 }
