### 2024-09-15 #7031 @cproctor
HI @SamHames, thank you for the thoughtful feedback, both on the software and on the argument. Thanks again to @cmaimone for raising issues earlier which would surely have gotten in your way as well.

I have addressed all the open issues in v1.5.4, and closed the issues--please let me know if you run into additional problems, or feel that my solutions are suboptimal. I have also revised the paper in response to your feedback; here's a summary of changes:

>     1. A background section or paragraph that introduces qualitative research for a non-qualitative audience (easier said than done, I know, but the statement of need assumes a lot of knowledge).

I agree, a background section is an improvement. I have added background on QDA and on CT, drawing largely on the documentation as you suggest.

>     2. Incorporation of the differences between qualitative data analysis in quantitative/mixed projects and pure qualitative projects. I find Kidder and Fine's (1987) big Q, little q qualitative distinction is a useful reference to direct people to.

I appreciate your sharing Kidder and Fine (1987)--I had not seen this before, and agree that it is a helpful distinction. I tried to incorporate little-q/big-Q with the existing framing, which distinguishes QDA as a *method* from qualitative/quantitative/mixed *research designs*.

>     3. A further expansion of what this means: "they are generally designed to protect users from complexity rather than providing affordances for engaging with complexity via algorithms and data structures". In particular some appropriate citation or situation within the literature. In many respects I'd say that the approach taken in qc also "protects users from complexity" by simply not engaging with many of the complexities that these packages routinely handle (such as annotation of arbitrary regions of text, preservation of rich document structure, multimodal annotation etc). qc gives up certain kinds of complexity to enable other kinds of engagement with textual qualitative data - it occupies a new and different space with different embedded values rather than a "better" space (and this is good!).

I have substantially reworked this section, including backing off from the specific claim you point out--I don't think it's my main point here to argue about the extent to which`qc` and conventional QDA abstract away complexity in their interfaces. (I have a related point in the documentation around how best to deal with leaky abstractions, which I'd also like to explore further--somewhere else.) Instead, I have tried to center the statement of need: users who would be ready to use CT in their QDA, but aren't well-served by the affordances of existing QDA tools.
