### 2024-09-16 #7031 @cproctor
Ah, I'm sorry--a careless mistake. I didn't have the latest version active when I ran the tests before deployment. This was an error introduced while correcting another issue.

This is fixed in 1.5.6; all tests passing and I have confirmed that the commands in the vignette work as expected.
