### 2024-09-17 cproctor/qualitative-coding#68 @cmaimone
anonymize and overlapping names
When the anonymize function creates anonymized versions of the documents, it does not handle names that may have overlap in them correctly.  For example, in my corpus, it identified:
Christina: Person_1
Christina M: Person_3

In the anonymized version of the document with "Christina M" in it, it replaced just Christina: "Person_1 M..."
