### 2024-09-17 #7031 @cmaimone
I've posted https://github.com/cproctor/qualitative-coding/issues/68 but it's more of a bug report than an issue for the JOSS review.  I was able to run through everything now.  I still ran into issues when I did anything unexpected (for example, accidentally edited a source document when coding), so I would encourage a reconsideration of the workflow.  But I think it meets the JOSS requirements at this point.
