### 2024-09-17 cproctor/qualitative-coding#68 @cproctor
This is functioning as designed, although I can see how in this situation the results might be surprising. It's a simple find-and-replace algorithm, iterating through the terms in the key file. So "Christina" got replaced with "Person_1" and then "Christina M" didn't match.

The most straightforward way of getting expected results using the current system would be to move "Christina M" above "Christina." This situation *could* be handled by organizing keys into a graph of substrings, and replacing the longest key strings first. I want to think about whether this might have other, more confusing side effects.

I'll either add a note to the documentation cautioning users about this, or will adopt the substring strategy.
