### 2024-09-18 377a86b
Bump to 1.6.0. Anonymize improved, with tests
