### 2024-09-18 ecf7e4e
Update version in docs
