### 2024-09-18 cproctor/qualitative-coding#68 @cproctor
I agree that the issue pointed out here should not happen, and have updated the command (with tests) in 1.6.0, the JOSS release version. Thanks!
