### 2024-09-18 #7031 @cproctor
[X] Double check authors and affiliations (including ORCIDs)
[X] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. **[Release v1.6.0](https://github.com/cproctor/qualitative-coding/releases/tag/v1.6.0)**
[X] Archive the release on Zenodo/figshare/etc and post the DOI here. **Archived release: [10.5281/zenodo.13784403](https://doi.org/10.5281/zenodo.13784403)**
[X] Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
[X] Make sure that the license listed for the archive is the same as the software license.

My post-review author checklist is complete. Thanks!
