### 2024-09-29 #7031 @ymzayek
Hi @cproctor I think the paper is almost ready. Can you just fix these minor typos?

L39: Atlas.TI -> ATLAS.ti
L39: MaxQDA -> MAXQDA
L42 wht -> with
