+++ paper.bib
@@ -12,6 +12,7 @@
   title = {Augmenting Human Intellect: A Conceptual Framework},
   author = {Engelbart, Douglas C},
   year = {1962},
+  doi = {10.21236/ad0289565},
   pages = {64--90}
 }
 @techreport{healy2020,
