### 2024-10-08 #7031 @crvernon
:wave: @cproctor - thank you for the well written submission which makes my job much easier.  I just have the following few items for you to address before I move to accept this work for publication:

**In the paper:**
- [ ] LINE 66:  The "d" in "Delphi" should be capitalized.  You can do this in your bib file by wrapping any characters you wish to maintain the formatting of in curly brackets.
- [ ] LINE 76:  The first "d" in "dedoose" should be capitalized.
- [ ] LINE 79:  "Atlas. Ti" should have no space and the "Ti" should both be lower case.

These changes do not require an additional release.  Please let me know when you have addressed them.  Thanks!
