### 2024-10-13 77c8e31
Add bibtex brackets to enforce capitalization in titles
