+++ paper.bib
@@ -51,7 +51,7 @@
   pages = {1902--1903}
 }
 @article{ogbeifun2016delphi,
-  title={The Delphi technique: A credible research methodology},
+  title={The {Delphi} technique: A credible research methodology},
   author={Ogbeifun, E and Agwa-Ejon, J and Mbohwa, Charles and Pretorius, JH},
   year={2016}
 }
@@ -84,7 +84,7 @@
   keywords = {computational thinking,equity,k12 cs}
 }
 @book{salmona2019qualitative,
-  title={Qualitative and mixed methods data analysis using Dedoose: A practical approach for research across the social sciences},
+  title={Qualitative and mixed methods data analysis using {Dedoose}: A practical approach for research across the social sciences},
   author={Salmona, Michelle and Lieber, Eli and Kaczynski, Dan},
   year={2019},
   publisher={Sage Publications}
@@ -96,7 +96,7 @@
   year = {1998}
 }
 @article{smit2002atlas,
-  title={Atlas. ti for qualitative data analysis},
+  title={{ATLAS.ti} for qualitative data analysis},
   author={Smit, Brigitte},
   journal={Perspectives in education},
   volume={20},
