### 2024-10-13 #7031 @cproctor
@crvernon thank you for your feedback and for catching these errors. I have pushed a fix and confirmed that the proof looks correct.
